almost done
to many changes to describe lol

+++ commands/MIRvote.js
@@ -13,11 +13,13 @@ module.exports = {
             this.end(message);
     },
     end(message) {
-        if (DATA.state() !== DATA.IN_MIR_VOTE) {
-            message.author.send("this command can be executed only after [!village vote start]");
+        
+        if (message && DATA.state() !== DATA.IN_MIR_VOTE ) {
+            DATA.narrator().send("this command can be executed only after [!village vote start]");
             message.delete();
             return;
         }
+
         DATA.setState(DATA.BEFORE_FIRST_NIGHT);
 
         let MIR;
@@ -36,14 +38,14 @@ module.exports = {
 
         });
 
-        let MIRAvatar = message.guild.members.cache.find((v)=> v.id === MIR.id).user.avatarURL();
+        //let MIRAvatar = message.guild.members.cache.find((v)=> v.id === MIR.id).user.avatarURL();
 
         const MIRDeclaration = new Discord.MessageEmbed()
             .setColor('#0099ff')
             .setTitle(MIR.tag)
             .setAuthor('Won elections for MIR')
             .setDescription('won vote by: '+MIR.candidate.count.length+' vote')
-            .setThumbnail(MIRAvatar)
+            .setThumbnail(MIR.avatarURL())
             .setTimestamp();
 
         DATA.messageChannel().send(MIRDeclaration);
@@ -51,32 +53,21 @@ module.exports = {
 
     },
     start(message) {
-        if (DATA.state() !== DATA.IN_ELECTION) {
+        if (message && DATA.state() !== DATA.IN_ELECTION) {
             message.author.send("the next command must be [!village vote start]");
             message.delete();
             return;
         }
         DATA.setState(DATA.IN_MIR_VOTE);
 
-        let nar = DATA.narrator();
-
-        if (nar.id !== message.author.id) {
+        if (DATA.narrator().id !== message.author.id) {
             message.delete();
             message.author.send('only narrator can execute this command');
             return;
         }
 
+        DATA.voteMessage('Vote started please react for one of the candidate :',DATA.messageChannel(), p=>{ p.candidate.on } );
 
-        DATA.messageChannel().send('Vote started please react for one of the candidate :');
-
-        DATA.players().forEach(async (player, i) => {
-
-            if (!player.candidate.on) return;
-
-            let playerVotingMessage = await DATA.messageChannel().send(`${i}. <@${player.id}>`);
-
-            playerVotingMessage.react('👍');
-        });
     }
 };
